Renamed Navbar component to index.tsx inside its folder.
